Arrumei a organização do git init e removi imports desnecessários
