+++ src/app/__init__.py
@@ -1,44 +1,69 @@
 from flask import Flask
-#Flask Security
-from flask_security.core import Security
-from flask_security import SQLAlchemyUserDatastore, Security
-#Flask SQLALCHEMY
+# Flask SQLALCHEMY
 from flask_sqlalchemy import SQLAlchemy
-#Flask Login
+# Flask Login
 from flask_login import LoginManager
-#Flask Formularios (WTForms)
-from wtforms import StringField, PasswordField, SubmitField
-from wtforms.validators import InputRequired, Length, ValidationError
-#Flask Admin
+# Flask Admin
 from flask_admin import Admin
 from flask_admin.contrib.sqla import ModelView
-#Flask Bcrypt
 from flask_bcrypt import Bcrypt
-#App config
-app = Flask(__name__)
-#App config SQLALCHEMY
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database/fatec.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-#App config Chaves de Segurança
-app.config['SECRET_KEY'] = 'Secreto'
-app.config['SECURITY_PASSWORD_SALT'] = 'muito seguro'
-app.config['SECURITY_PASSWORD_HASH'] = 'pbkdf2_sha512'
-#Banco de Dados Config
-db = SQLAlchemy(app)
-#App Bcrypt
-bcrypt = Bcrypt(app)
-# Flask Admin Config
-from .database.models import *
-admin = Admin(app, name='FATEC SJC')
-admin.add_view(ModelView(Postagem, db.session))
-admin.add_view(ModelView(User, db.session))
-#Login Maneger
-login_manager = LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = "auth.login"
-login_manager.login_message = 'Você deve logar-se para acessar essa página'
-#Importação de Rotas atraves da BluePrint
-from .auth import routes as auth_blueprint
-app.register_blueprint(auth_blueprint)
-from .view import routes as view_blueprint
-app.register_blueprint(view_blueprint)
+
+# Path
+from os import path
+
+# Banco de dados config
+db = SQLAlchemy()
+DB_NAME = "fatec.db"
+
+# Bcrypt
+bcrypt = Bcrypt()
+
+# App config
+
+
+def create_app():
+    app = Flask(__name__)
+
+    # App config Chaves de Segurança
+    app.config['SECRET_KEY'] = 'Secreto'
+    app.config['SECURITY_PASSWORD_SALT'] = 'muito seguro'
+    app.config['SECURITY_PASSWORD_HASH'] = 'pbkdf2_sha512'
+
+    # App config SQLALCHEMY
+    app.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///database/{DB_NAME}'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
+    # Init
+    db.init_app(app)
+    bcrypt.init_app(app)
+
+    # Importação de Rotas atraves da BluePrint
+    from .auth import routes as auth_blueprint
+    from .view import routes as view_blueprint
+
+    app.register_blueprint(auth_blueprint)
+    app.register_blueprint(view_blueprint)
+
+    # Flask Admin Config
+    from .database.models import Postagem, User
+    admin = Admin(app, name='FATEC SJC')
+    admin.add_view(ModelView(Postagem, db.session))
+    admin.add_view(ModelView(User, db.session))
+
+    # Login Maneger
+    login_manager = LoginManager()
+    login_manager.login_view = "auth.login"
+    login_manager.login_message = 'Você deve logar-se para acessar essa página'
+    login_manager.init_app(app)
+
+    @login_manager.user_loader
+    def load_user(user_id):
+        return User.query.get(int(user_id))
+
+    return app
+
+
+def create_database(app):
+    if not path.exists('app/database/' + DB_NAME):
+        db.create_all(app=app)
+        print('Created Database!')
