+++ src/app/templates/base-login.html
@@ -76,9 +76,11 @@
   </footer>
 
 </body>
+<!--
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.mask/1.14.16/jquery.mask.min.js"></script>
-<script src="{{url_for('static', filename='js/main.js')}}"></script>
+<script src="{{url_for('static', filename='js/main.js')}}"></script>-->
+<script src="https://unpkg.com/imask"></script>
 <script src="{{url_for('static', filename='js/mask.js')}}"></script>
 
 </html>
